+++ src/functions/arithmetic/subtract.js
@@ -12,12 +12,21 @@ class LimeFunctionSubtract extends LimeFunction {
 
     // Binary operation
     this.operations.b = [
+      'cb(int>rat,rat)',
+      'cb(rat,int>rat)',
       'eb(int,int)',
+      'eb(rat,rat)',
     ];
 
     // Algorithms
     this.algorithms.set('b(int,int)', (step) => {
       step.bs(this.lime.build('integer')(step.left.value - step.right.value));
+    });
+    this.algorithms.set('b(rat,rat)', (step) => {
+      step.bs(this.lime.build('rational')(
+        this.lime.build('integer')(step.left.value.n * step.right.value.d - step.right.value.n * step.left.value.d),
+        this.lime.build('integer')(step.left.value.d * step.right.value.d),
+      ));
     });
   }
 }