+++ docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ### v0.1.0
 - Set up engine framework, bundling system using Rollup and testing system using Jest
 - Set up prompt and web examples
-- Added expression: `Integer`
+- Added expression: `Integer`, `Rational`
 - Added functions:
-  - Arithmetic: `Add`, `Multiply`, `Subtract`
+  - Arithmetic: `Add`, `Divide`, `Multiply`, `Subtract`
 - Added tests: `Banner`, `Bundle`, `Error`, `Expressions`, `Functions`, `Module`, `Types`